Added git-flow, git-completion, tree, and Ant 1.10.0.

+++ chef/bento/scripts/oracle/bootstrap_ol7_ansible.sh
@@ -1,14 +1,18 @@
 #!/bin/sh -eux
 # install ansible on oracle linux 7.x.
 
-# create temporary scripts directory. ------------------------------------------
-mkdir -p /tmp/scripts/oracle
-cd /tmp/scripts/oracle
+# install epel repository if needed. -------------------------------------------
+epel_repo="/etc/yum.repos.d/epel.repo"
+if [ ! -f "$epel_repo" ]; then
+  # create temporary scripts directory.
+  mkdir -p /tmp/scripts/oracle
+  cd /tmp/scripts/oracle
 
-# install the epel repository. -------------------------------------------------
-wget --no-verbose https://dl.fedoraproject.org/pub/epel/epel-release-latest-7.noarch.rpm
-yum repolist
-yum -y install epel-release-latest-7.noarch.rpm
+  # install the epel repository.
+  wget --no-verbose https://dl.fedoraproject.org/pub/epel/epel-release-latest-7.noarch.rpm
+  yum repolist
+  yum -y install epel-release-latest-7.noarch.rpm
+fi
 
 # install ansible. -------------------------------------------------------------
 yum -y install ansible
